+++ src/EstadoLiquidacion.java
@@ -1,9 +1,9 @@
 public class EstadoLiquidacion implements EstadoDePrenda{
 
-    private final float porcentajeLiquidacion = 0.5f;
+    private final Double porcentajeLiquidacion = 0.5;
 
     @Override
-    public int descuento(int precioInicial) {
-        return (int) (precioInicial * porcentajeLiquidacion);
+    public Double descuento(Double precioInicial) {
+        return precioInicial * porcentajeLiquidacion;
     }
 }
