+++ src/EstadoNuevo.java
@@ -1,6 +1,6 @@
 public class EstadoNuevo implements EstadoDePrenda {
     @Override
-    public int descuento(int precioInicial) {
-        return 0;
+    public Double descuento(Double precioInicial) {
+        return 0.0;
     }
 }
